+++ lib/generics/include/list.h
@@ -1,20 +1,22 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
 
 #include "defines.h"
 
 /* node object */
 struct node {
   unsigned char *data;
-  unsigned long long data_size;
+  size_t data_size;
   struct node *next;
   struct node *prev;
 };
 
 /* doubly linked list object */
 struct list {
-  unsigned long long size;  // can never exceeds LLONG_MAX
+  size_t size;  // can never exceeds LLONG_MAX
   struct node *head;
   struct node *tail;
 };
@@ -32,29 +34,29 @@ void list_destroy(struct list *list, void (*destroy)(void *data));
 
 /* returns the number of elements on the list. avoid acceessing list::size
  * directly. use this method instead */
-unsigned long long list_size(struct list *list);
+size_t list_size(struct list *list);
 
 /* returns whether vector is emtpy or not. if list is NULL - returns true */
 bool list_empty(struct list *list);
 
 /* add a node to the start of the list. returns true on success, false otherwise
  */
-bool list_prepend(struct list *list, const void *data, unsigned long long data_size);
+bool list_prepend(struct list *list, const void *data, size_t data_size);
 
 /* add a node to the end of the list. returns true on success, false otherwise
  */
-bool list_append(struct list *list, const void *data, unsigned long long data_size);
+bool list_append(struct list *list, const void *data, size_t data_size);
 
 /* constructs and inserts a node at position pos. the position is calculated
  * from the head of the list. returns true on success, false otherwise */
-bool list_insert_at(struct list *list, const void *data, unsigned long long data_size, unsigned long long pos);
+bool list_insert_at(struct list *list, const void *data, size_t data_size, size_t pos);
 
 /* constructs and inserts a node at a position determined by the comparator.
  * inserts the new node at the first location where new_node::data >
  * another_node::data */
 bool list_insert_priority(struct list *list,
                           const void *data,
-                          unsigned long long data_size,
+                          size_t data_size,
                           int (*cmpr)(const void *, const void *));
 
 /* returns a pointer to the first element on the list. returns NULL on failure
@@ -65,7 +67,7 @@ void *list_peek_first(struct list *list);
 void *list_peek_last(struct list *list);
 
 /* returns a pointer to the element at position pos. returns NULL on failure. */
-void *list_at(struct list *list, unsigned long long pos);
+void *list_at(struct list *list, size_t pos);
 
 /* removes the first node from the list. returns a pointer to the removed
  * element on success (which has to be free'd), NULL on failure */
@@ -78,24 +80,24 @@ void *list_remove_last(struct list *list);
 /* removes the node at position pos. the position is caluclated from the head of
  * the list. returns a pointer to the removed element on success (which has to
  * be free'd), NULL otherwise */
-void *list_remove_at(struct list *list, unsigned long long pos);
+void *list_remove_at(struct list *list, size_t pos);
 
 /* finds the first occurence of data and returns its index. the index is
  * calculated from the list's head. returns a positive number as the index on
  * success, negative number on failure */
-long long list_index_of(struct list *list, const void *data, int (*cmpr)(const void *, const void *));
+intmax_t list_index_of(struct list *list, const void *data, int (*cmpr)(const void *, const void *));
 
 /* replaces an element at position pos. the position is calculated from the
  * list's head. returns a pointer to the replcaed data on success (which has to
  * be free'd), NULL otherwise */
-void *list_replace_at(struct list *list, const void *data, unsigned long long data_size, unsigned long long pos);
+void *list_replace_at(struct list *list, const void *data, size_t data_size, size_t pos);
 
 /* replaces the first occurence of old_data with new_data. returns a pointer to
  * old_data on success (which has to be free'd), NULL otherwise */
 void *list_replace(struct list *list,
                    const void *old_data,
                    const void *new_data,
-                   unsigned long long new_data_size,
+                   size_t new_data_size,
                    int (*cmpr)(const void *, const void *));
 
 /* sorts the list */
